refactor: 無駄なファイル切り出し削除
